requirements.txt
absl-py==2.1.0
aiohappyeyeballs==2.4.3
aiohttp==3.11.8
aiosignal==1.3.1
astunparse==1.6.3
async-timeout==5.0.1
attrs==24.2.0
breadability==0.1.20
certifi==2024.8.30
chardet==5.2.0
charset-normalizer==3.4.0
click==8.1.7
datasets==3.1.0
dill==0.3.8
docopt==0.6.2
filelock==3.16.1
flatbuffers==24.3.25
frozenlist==1.5.0
fsspec==2024.9.0
gast==0.6.0
google-pasta==0.2.0
h5py==3.12.1
hdbscan==0.8.40
huggingface-hub==0.26.3
idna==3.10
Jinja2==3.1.4
joblib==1.4.2
keras==3.7.0
libclang==18.1.1
llvmlite==0.43.0
lxml==5.3.0
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
ml-dtypes==0.3.2
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.16
namex==0.0.8
networkx==3.4.2
nltk==3.9.1
numba==0.60.0
numpy==1.26.0
opt_einsum==3.4.0
optree==0.13.1
packaging==24.2
pandas==1.5.3
pillow==11.0.0
plotly==5.16.1
propcache==0.2.0
protobuf==4.25.5
pyarrow==18.1.0
pycountry==24.6.1
Pygments==2.18.0
pynndescent==0.5.13
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
rich==13.9.4
safetensors==0.4.5
scikit-learn==1.2.2
scipy==1.10.1
sentence-transformers==2.2.2
sentencepiece==0.2.0
six==1.16.0
sumy==0.11.0
sympy==1.13.3
tenacity==9.0.0
tensorboard==2.16.2
tensorboard-data-server==0.7.2
termcolor==2.5.0
threadpoolctl==3.5.0
tokenizers==0.20.3
torch==2.0.1
torchvision==0.15.2
tqdm==4.67.1
typing_extensions==4.12.2
umap-learn==0.5.3
urllib3==2.2.3
Werkzeug==3.1.3
wrapt==1.17.0
xxhash==3.5.0
yarl==1.18.0
